+++ src/backend/resolvers/mutations/importWorkbook/utils/validate/tests/businessObject/businessObject.test.js
@@ -13,6 +13,14 @@ const {
   BO_OUTPUT_2_errors,
 } = require('./mockData/output2')
 
+const {
+  BO_INPUT_3_dataAndSkippedRows,
+  BO_INPUT_3_sheetConfig,
+} = require('./mockData/input3')
+const BO_OUTPUT_3_data = require('./mockData/output3')
+
+const BO_OUTPUT_4_errors = require('./mockData/output4')
+
 describe('Test business object validation', () => {
   test(`- Valid data with business object validation passes validation
     - oneOf disregarded when businessObj ref exists`, async () => {
@@ -52,5 +60,43 @@ describe('Test business object validation', () => {
     expect(valid).toEqual(false)
     expect(errors).toStrictEqual(BO_OUTPUT_2_errors)
     expect(data).toStrictEqual(BO_OUTPUT_2_data)
+  })
+
+  test(`Data with blank cells pass when businessObjRef.allowBlankValues is true`, async () => {
+    const { result, skippedRows } = BO_INPUT_3_dataAndSkippedRows
+
+    const {
+      valid,
+      errors,
+      data,
+    } = await validate({
+      data: result,
+      skippedRows,
+      sheetConfig: BO_INPUT_3_sheetConfig,
+      db: BO_INPUT_1_mockDb,
+    })
+
+    expect(valid).toEqual(true)
+    expect(errors).toStrictEqual([])
+    expect(data).toStrictEqual(BO_OUTPUT_3_data)
+  })
+
+  test(`Data with blank cells fail when businessObjRef.allowBlankValues is false`, async () => {
+    const { result, skippedRows } = BO_INPUT_3_dataAndSkippedRows
+
+    const {
+      valid,
+      errors,
+      data,
+    } = await validate({
+      data: result,
+      skippedRows,
+      sheetConfig: BO_INPUT_1_sheetConfig,
+      db: BO_INPUT_1_mockDb,
+    })
+
+    expect(valid).toEqual(false)
+    expect(errors).toStrictEqual(BO_OUTPUT_4_errors)
+    expect(data).toStrictEqual(BO_OUTPUT_3_data)
   })
 })