+++ src/backend/resolvers/mutations/importWorkbook/utils/validate/tests/businessObject/mockData/input3/index.js
@@ -1,0 +1,7 @@
+const BO_INPUT_3_dataAndSkippedRows = require('./data-and-skipped-rows')
+const BO_INPUT_3_sheetConfig = require('./sheet-config')
+
+module.exports = {
+  BO_INPUT_3_dataAndSkippedRows,
+  BO_INPUT_3_sheetConfig,
+}